Move all of the non-batching code into `streams.processing`.
