+++ snuba/consumers/consumer_builder.py
@@ -11,7 +11,7 @@ from snuba.snapshots import SnapshotId
 from snuba.stateful_consumer.control_protocol import TransactionData
 from snuba.utils.metrics.backends.wrapper import MetricsWrapper
 from snuba.utils.retries import BasicRetryPolicy, RetryPolicy, constant_delay
-from snuba.utils.streams.batching import BatchingConsumer, BatchProcessorFactory
+from snuba.utils.streams.batching import BatchProcessorFactory
 from snuba.utils.streams.kafka import (
     KafkaConsumer,
     KafkaConsumerWithCommitLog,
@@ -19,6 +19,7 @@ from snuba.utils.streams.kafka import (
     TransportError,
     build_kafka_consumer_configuration,
 )
+from snuba.utils.streams.processing import BatchingConsumer
 from snuba.utils.streams.types import Topic
