+++ snuba/stateful_consumer/states/catching_up.py
@@ -3,8 +3,8 @@ from typing import Optional, Tuple
 
 from snuba.consumers.consumer_builder import ConsumerBuilder
 from snuba.stateful_consumer import ConsumerStateData, ConsumerStateCompletionEvent
-from snuba.utils.streams.batching import BatchingConsumer
 from snuba.utils.streams.kafka import KafkaPayload
+from snuba.utils.streams.processing import BatchingConsumer
 from snuba.utils.state_machine import State
 
 logger = logging.getLogger("snuba.snapshot-catchup")
